make retry / retryTImeout reusable
